fix(ci): python-full-coverage — split three files over the 1500-line maintainability cap
The merge from development pushed three first-party files past the
MAX_FIRST_PARTY_FILE_LINES=1500 guardrail, failing
test_first_party_code_files_stay_under_line_limit (the CI summary truncated
the failure repr to a single file, but three were over).

Decompose each by relocation only (no logic changes):
- src/awf/service/gc.py (1534): extract leaf result dataclasses into a new
  sibling awf/service/gc_results.py, re-exported from gc.py to keep the
  awf.service.gc.<Name> import surface intact.
- tests/unit/service/test_provider_readiness_parts/test_provider_readiness_part_002.py
  (1558): move trailing redaction/fallback tests into a new part_004 sibling.
- tests/unit/node/test_claude_auth_overlay.py (1882): split into a
  test_claude_auth_overlay_parts/ package (part_001 + part_002), matching the
  repo's existing parts convention.
